Change the definition of entropy in bigram_lattice.py
The code followed the definition of expected value, and entropy
Sum_<all bigrams> P(bigram) log2(1/P(bigram))
Since the calculation isn't over all possible bigrams, but only those
which occur in the document under examination, I changed it to
(Sum_<bigrams in document> log2(1/P(bigram))) /Count_<bigrams in document>,
which has the effect of approximating the probability of the the
bigrams in the document with their frequencies.  This estimate doesn't
include any extra smoothing to reflect the fact that the document we are
testing is only representing the universe of similar documents, so
any probabilities we see in the document under test are even less
precise approximations than the probabilities we have already tabulated.

(In the paper I claim that the two universes of dialect
transcriptions and newspaper articles have acceptable cross-entropy
under my hierarchical backoff scheme, and this is interesting because
we have so few dialect transcriptions.)